backend/requirements.txt
Django>=6.0
djangorestframework
django-model-utils
psycopg2-binary
dj-database-url
python-dotenv
django-environ
django-cors-headers
gunicorn
whitenoise
